This commit was manufactured by cvs2svn to create tag 'ANT_15_B2'.
git-svn-id: https://svn.apache.org/repos/asf/ant/core/tags/ANT_15_B2@272825 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/org/apache/tools/ant/taskdefs/Move.java
@@ -306,7 +306,7 @@ public class Move extends Copy {
                     }
                 }
 
-                if (destFile.exists()) {
+                if (destFile.exists() && destFile.isFile()) {
                     if (!destFile.delete()) {
                         throw new BuildException("Unable to remove existing "
                                                  + "file " + destFile);
